+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-# 1.0.2
+# Changelog
 
-Initial release
+## 1.1.0
+
+- Added a `matches(shortcut: string)` to `ShortcutEvent`. This can be used to check a shortcut against a list of aliases. This is now the prefered way checking a shortcut.
+- `ShortcutEvent.shortcut` and `ShortcutEvent.keys` now use PascalCase.
+- Improved `README.md`
+
+## 1.0.0
+
+- Initial release